Fix parsing of out-of-range PCIe lane count values
It seems that, depending on the firmware version, out-of-range values
for the PCIe lane count may be handled differently. Sometimes an
out-of-range value will cause the firmware to set the number of lanes to
the maximum supported by the chip. Other times, it will choose a totally
different value, like one lane instead of four on the ASM2364. From what
I can tell from the testing I've done, the most reliable way to tell the
firmware to use the maximum number of lanes is to specifically set the
lane count to two lanes for the ASM2362 and four lanes for the ASM2364.